h is a floating point number with the new instances nomenclature

+++ src/network/graph_info.h
@@ -5,7 +5,7 @@
 
 struct graph_info {
     int n;
-    int h;
+    double h;
     double k;
     int capacity;
